analyst_generix/requirements.txt
crewai>=0.11.0
langchain>=0.1.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
openai>=1.3.0 